+++ Clima/app/src/main/java/br/com/rbarrelo/clima/openweathermap/pojo/OpenWeatherMain.java
@@ -6,8 +6,8 @@ package br.com.rbarrelo.clima.openweathermap.pojo;
 public class OpenWeatherMain {
 
     private double temp;
-    private int pressure;
-    private int humidity;
+    private double pressure;
+    private double humidity;
     private double temp_min;
     private double temp_max;
 
@@ -19,19 +19,19 @@ public class OpenWeatherMain {
         this.temp = temp;
     }
 
-    public int getPressure() {
+    public double getPressure() {
         return pressure;
     }
 
-    public void setPressure(int pressure) {
+    public void setPressure(double pressure) {
         this.pressure = pressure;
     }
 
-    public int getHumidity() {
+    public double getHumidity() {
         return humidity;
     }
 
-    public void setHumidity(int humidity) {
+    public void setHumidity(double humidity) {
         this.humidity = humidity;
     }
